Add --rmout option to mandelbrot.
When estimating distance, weird artifacts appear in places, but disappear
when merely changing the number of pixels in the image.

The --rmout option addresses this by cropping statistical outliers.  Some subtle
color variations may disappear, especially in the brightest areas.

--rmout takes an optional argument, --rmout=MULTIPLE, where MULTIPLE is a multiple
of the standard deviation to crop things off at.  If not used, this is 3.0, since
that's generally considered the outlier threshold of normal distribution.  But since
these images do not necessarily follow normal distribution, the paramter may need
to be tuned to taste.