Add --rmout option to buddhabrot1.
This is the same as the --rmout option to mandelbrot, except that it serves
a different purpose, since buddhabrot1 does not do distance estimates (and
therefore statistical-outlier artifacts are not a concern).  But since the
brightest parts of a buddhabrot are insanely brighter than most of the picture,
this serves as just an alternative to histogram equalization.  Experimentally
it appears slightly less grainy than using the histogram.  But for more polished
results, it's best to filter an untouched image in Photoshop than to try to
tweak it on the command line.