Interleave the rows used by each thread in mbrot2.
Previous method naively gave each thread its own big chunk of the image
to render, from top to bottom.  But the image's complexity is not evenly
distributed like that.  It's possible for one thread to still be doing
ninety percent of the work.  So have each thread's row selection be
interleaved with each other; it still won't be perfect, but it should be
a dramatic improvement on images where, say, the middle of a picture has
the greatest complexity.