Completely rewrite pxbuf and the programs that use it, except for the ones in examples/.
Also, since mbrot2 is so well on its way, deleted mandelbrot/.

This greatly imrpoves the --rmout option for bbrot2 (though not so much
for mbrot2), fixes the problem with mbrot2 outputs that fade when scaling
the -w and -h options, and fixes the problem wherein normalizing and
equalization were re-balancing the RGB channels.  For legacy's sake, this
rebalancing still occurs by default, unless the --linked option is used.